Changes Sequelize's configs
- In the hope to remove double quotes in the names of database tables,
and make things more consistent, I tried to configure Sequelize to not
use camelCased names in the auto generated columns. For instance, I
wanted createdAt to be created_at. It seems to be a more common apporach
when using Postrges, but it is a pain to set up in Sequelize.